backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg[binary]==3.2.3
psycopg-pool==3.2.4
pydantic-settings==2.6.1
httpx==0.28.1